// src/actions/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Fri, 14 Jul 2023 19:22:17 GMT
 */

/**
 * @enum
 * @readonly
 * @description MsigWorlds Action Names
 */
export enum MsigWorldsActionName {
    Approve = 'approve',
    Blockaction = 'blockaction',
    Cancel = 'cancel',
    Checkauth = 'checkauth',
    Cleanup = 'cleanup',
    Deny = 'deny',
    Exec = 'exec',
    Invalidate = 'invalidate',
    Propose = 'propose',
    Unapprove = 'unapprove',
};